Removed redundant setting of form_bond parameters
These don't need to be set to false because they're erased by the
launch file, and if they aren't found, they're considered false.  When
they are set like they are now, they can override other parts of the
launch file setting them to true before this parameter file is loaded.